requirements.txt
aiogram==3.16.0
python-dotenv==1.0.0
mcp==1.0.0
pydantic>=2.4.1,<2.11
aiohttp==3.9.1
python-dateutil==2.8.2
pytz==2023.3
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.110.0